Portfolio formatting. Changes on p_analysis data type.

+++ database/migrations/2019_03_15_230245_create_p_analyses_table.php
@@ -16,11 +16,11 @@ class CreatePAnalysesTable extends Migration
         Schema::create('p_analyses', function (Blueprint $table) {
             $table->increments('id');
             $table->string('class');
-            $table->decimal('n_of_months', 25, 10)->default(0);
+            $table->integer('n_of_months')->default(0);
             $table->decimal('ltm', 25, 10)->default(0);
             $table->decimal('overall', 25, 10)->default(0);
             $table->decimal('sharpe_ratio', 25, 10)->default(0);
-            $table->decimal('wml', 25, 10)->default(0);
+            $table->integer('wml')->default(0);
             $table->decimal('st_deviation', 25, 10)->default(0);
               
         });
